# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi berhasil, user dibuat (password tidak pernah dikembalikan)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Registrasi berhasil, user dibuat (password tidak pernah dikembalikan)"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/User"
                        },
                        {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      ],
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "summary": "Register user baru",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Menyetel cookie `access_token` dan `refresh_token` (HttpOnly). Access token dan refreshToken juga dikembalikan di response.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil, cookie access_token dan refresh_token diset",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Success",
                    "data": {
                      "message": "Login successful",
                      "user": {
                        "id": 1,
                        "email": "user@example.com",
                        "role": "USER"
                      },
                      "refreshToken": "opaque-refresh-token"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Login dan set access token serta refresh token ke cookie HttpOnly",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Membaca refreshToken dari body atau cookie `refresh_token`. Mengembalikan accessToken + refreshToken baru dan memperbarui cookie HttpOnly.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rotasi berhasil, cookie access_token dan refresh_token diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Success",
                    "data": {
                      "accessToken": "jwt.access.token",
                      "refreshToken": "opaque-refresh-token-baru"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Rotasi refresh token dengan grace period",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesi revoke, cookie access_token dan refresh_token dihapus",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Success",
                    "data": {
                      "message": "Logged out"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Logout: revoke refresh token sesi dan hapus cookie",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    }
                  },
                  "required": [
                    "success",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Lihat profil user yang sedang login",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Library Digital API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Alternatif untuk cookie: tempel access_token hasil login."
      },
      "access_token": {
        "type": "http",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly cookie access token, diset otomatis oleh POST /auth/login."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "identification_number": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "role": {
            "enum": [
              "SUPER_ADMIN",
              "STAFF",
              "USER"
            ],
            "type": "string"
          },
          "category": {
            "enum": [
              "STUDENT",
              "LECTURER",
              "LIBRARY_STAFF",
              "PUBLIC"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "identification_number",
          "email",
          "full_name",
          "role",
          "category",
          "created_at",
          "updated_at"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "identification_number": {
            "type": "string",
            "example": "1234567890",
            "description": "Nomor identitas unik"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "pattern": "/(?=.*[a-zA-Z])(?=.*[0-9])/",
            "example": "Str0ng!P4ss#2024"
          },
          "full_name": {
            "type": "string",
            "example": "John Doe"
          },
          "category": {
            "enum": [
              "STUDENT",
              "LECTURER",
              "LIBRARY_STAFF",
              "PUBLIC"
            ],
            "type": "string",
            "example": "STUDENT"
          }
        },
        "required": [
          "identification_number",
          "email",
          "password",
          "full_name",
          "category"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "Str0ng!P4ss#2024"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "opaque-refresh-token-string",
            "description": "Refresh token yang dikembalikan saat login atau dikirim lewat cookie"
          }
        }
      },
...
    }
  }
}